requirements.txt
aiohttp==1.0.5
async-timeout==3.0.1
beautifulsoup4==4.8.1
chardet==3.0.4
discord.py==0.16.12
multidict==4.5.2
python-dotenv==0.10.3
soupsieve==1.9.5
urllib3==1.25.6
websockets==3.4
